Builds and runs on Windows, still a segfault though

+++ src/renderer.cpp
@@ -13,7 +13,7 @@
 GLuint Texture = 0;
     GLuint texture = 1;
 
-  GLdouble norm = 1 / sqrt( 3 );
+  GLdouble norm = 1 / sqrtf( 3 );
 
     GLdouble x = 0;
     GLdouble y = 0;
